+++ python/ray/serve/backend_worker.py
@@ -2,18 +2,17 @@ import asyncio
 import logging
 import traceback
 import inspect
-from typing import Union, Any, Callable, Type, Optional
+from typing import Union, Any, Callable, Type
 import time
 
 import starlette.responses
-from starlette.requests import Request
 
 import ray
 from ray.actor import ActorHandle
 from ray._private.async_compat import sync_to_async
 
-from ray.serve.utils import (ASGIHTTPSender, parse_request_item, _get_logger,
-                             import_attr)
+from ray.serve.http_util import ASGIHTTPSender
+from ray.serve.utils import (parse_request_item, _get_logger, import_attr)
 from ray.serve.exceptions import RayServeException
 from ray.util import metrics
 from ray.serve.config import BackendConfig
@@ -23,7 +22,6 @@ from ray.serve.constants import (
     BACKEND_RECONFIGURE_METHOD,
     DEFAULT_LATENCY_BUCKET_MS,
 )
-from ray.serve.http_util import make_startup_shutdown_hooks
 from ray.exceptions import RayTaskError
 
 logger = _get_logger()
@@ -66,7 +64,8 @@ def create_backend_replica(backend_def: Union[Callable, Type[Callable], str]):
             if is_function:
                 _callable = backend
             else:
-                _callable = backend(*init_args)
+                _callable = backend.__new__(backend)
+                await sync_to_async(_callable.__init__)(*init_args)
             # Setting the context again to update the servable_object.
             ray.serve.api._set_internal_replica_context(
                 backend_tag,
@@ -74,22 +73,10 @@ def create_backend_replica(backend_def: Union[Callable, Type[Callable], str]):
                 controller_name,
                 servable_object=_callable)
 
-            self.shutdown_hook: Optional[Callable] = None
-            if backend_config.internal_metadata.is_asgi_app:
-                app = _callable._serve_asgi_app
-                startup_hook, self.shutdown_hook = make_startup_shutdown_hooks(
-                    app)
-                await startup_hook()
-
             assert controller_name, "Must provide a valid controller_name"
             controller_handle = ray.get_actor(controller_name)
             self.backend = RayServeReplica(_callable, backend_config,
                                            is_function, controller_handle)
-
-        def __del__(self):
-            if hasattr(self, "shutdown_hook") and self.shutdown_hook:
-                asyncio.get_event_loop().run_until_complete(
-                    self.shutdown_hook())
 
         @ray.method(num_returns=2)
         async def handle_request(
@@ -246,20 +233,10 @@ class RayServeReplica:
         start = time.time()
         method_to_call = None
         try:
-            # TODO(simon): Split this section out when invoke_batch is removed.
-            if self.config.internal_metadata.is_asgi_app:
-                request: Request = args[0]
-                sender = ASGIHTTPSender()
-                await self.callable._serve_asgi_app(
-                    request.scope,
-                    request._receive,
-                    sender,
-                )
-                result = sender.build_starlette_response()
-            else:
-                method_to_call = sync_to_async(
-                    self.get_runner_method(request_item))
-                result = await method_to_call(*args, **kwargs)
+            method_to_call = sync_to_async(
+                self.get_runner_method(request_item))
+            result = await method_to_call(*args, **kwargs)
+
             result = await self.ensure_serializable_response(result)
             self.request_counter.inc()
         except Exception as e:
