+++ python/ray/serve/tests/test_backend_worker.py
@@ -6,7 +6,7 @@ import ray
 from ray.serve.backend_worker import create_backend_replica, wrap_to_ray_error
 from ray.serve.controller import TrafficPolicy
 from ray.serve.router import RequestMetadata, EndpointRouter
-from ray.serve.config import BackendConfig, BackendMetadata
+from ray.serve.config import BackendConfig
 from ray.serve.utils import get_random_letters
 
 pytestmark = pytest.mark.asyncio
@@ -157,8 +157,7 @@ async def test_task_runner_perform_async(serve_instance,
         await barrier.wait.remote()
         return "done!"
 
-    config = BackendConfig(
-        max_concurrent_queries=10, internal_metadata=BackendMetadata())
+    config = BackendConfig(max_concurrent_queries=10)
 
     worker, router = await add_servable_to_router(
         wait_and_go, *mock_controller_with_name, backend_config=config)
@@ -233,9 +232,7 @@ async def test_graceful_shutdown(serve_instance, mock_controller_with_name):
         KeepInflight,
         *mock_controller_with_name,
         backend_config=BackendConfig(
-            num_replicas=1,
-            internal_metadata=BackendMetadata(),
-            user_config={"release": False}))
+            num_replicas=1, user_config={"release": False}))
 
     query_param = make_request_param()
